Fixed a bug that was tiling cairo-dock.
I've also added functionality to skip tiling of windows that are splash
screens, or set to skip the taskbar/pager, etc. Of course, that is
assuming that the window sets the proper properties. (Case in point: the
Eclipse splash screen will tile since it doesn't set _NET_WM_WINDOW_TYPE
properly, but OpenOffice's splash screen will not tile.)